Refactoring energy data for server side processing.
