+++ components/charts/ChartEnergy.jsx
@@ -2,37 +2,9 @@ import { ResponsiveBar } from "@nivo/bar";
 import Loading from "../Loading";
 
 const ChartEnergy = ({ energyData }) => {
-  if (!energyData || !energyData.mainData) return <Loading />;
+  console.log("energyData: ", energyData);
 
-  const capThreshold = 1100000;
-  const processedData = energyData.mainData.reduce((acc, item) => {
-    const country = item["Country Name"];
-    const productionGWh =
-      parseFloat(
-        item[
-          "Renewable electricity output (GWh) [4.1.2_REN.ELECTRICITY.OUTPUT]"
-        ]
-      ) || 0;
-    let consumptionGWh =
-      (parseFloat(
-        item["Renewable energy consumption (TJ) [3.1_RE.CONSUMPTION]"]
-      ) || 0) * 0.277778;
-
-    if (country === "China" || country === "India") {
-      consumptionGWh = Math.min(consumptionGWh, capThreshold);
-    }
-
-    acc[country] = acc[country] || { production: 0, consumption: 0 };
-    acc[country].production += productionGWh;
-    acc[country].consumption += consumptionGWh;
-
-    return acc;
-  }, {});
-
-  const chartData = Object.entries(processedData).map(([country, data]) => ({
-    country,
-    ...data,
-  }));
+  if (!energyData) return <Loading />;
 
   const formatValue = (value, forTooltip = false) => {
     let formattedValue;
@@ -84,7 +56,7 @@ const ChartEnergy = ({ energyData }) => {
         <h2 className="chart-title">Renewable Energy (GWh)</h2>
         <h2 className="chart-title">Production and Consumption </h2>
         <ResponsiveBar
-          data={chartData}
+          data={energyData}
           keys={["production", "consumption"]}
           indexBy="country"
           margin={{ top: 15, right: 15, bottom: 115, left: 50 }}
